# Conflicts:
#	Assets/Art/Particles/VFX_Materials/VFX_FireSparks_AddBlend_02.mat
#	Assets/Game/Gameplay/Camera/FGOL/GameCamera.cs
#	Assets/Resources/Game/Dragons/PF_DragonBaby.prefab
#	Assets/Resources/Game/Dragons/PF_DragonBalrog.prefab
#	Assets/Resources/Game/Dragons/PF_DragonBug.prefab
#	Assets/Resources/Game/Dragons/PF_DragonChinese.prefab
#	Assets/Resources/Game/Dragons/PF_DragonClassic.prefab
#	Assets/Resources/Game/Dragons/PF_DragonCrocodile.prefab
#	Assets/Resources/Game/Dragons/PF_DragonDevil.prefab
#	Assets/Resources/Game/Dragons/PF_DragonFat.prefab
#	Assets/Resources/Game/Dragons/PF_DragonJawfrey.prefab
#	Assets/Resources/Game/Dragons/PF_DragonReptile.prefab
#	Assets/Resources/Game/Dragons/PF_DragonTitan.prefab


Former-commit-id: 39ec052fce7c9349c4bf4abd6369dd114cdc3a47

+++ Assets/Game/Model/MissionsAndEvents/Missions/Mission.cs
@@ -99,7 +99,7 @@ public class Mission {
 	/// <param name="_missionDef">Mission definition.</param>
 	/// <param name="_targetValue">Target value.</param>
 	/// <param name="_singleRun">Is it a single run mission?</param>
-	public void InitWithParams(DefinitionNode _missionDef, DefinitionNode _typeDef, float _targetValue, bool _singleRun) {
+	public void InitWithParams(DefinitionNode _missionDef, DefinitionNode _typeDef, long _targetValue, bool _singleRun) {
 		// Store definitions
 		m_def = _missionDef;
 		m_typeDef = _typeDef;
@@ -297,7 +297,7 @@ public class Mission {
 		InitWithParams(
 			missionDef,
 			DefinitionsManager.SharedInstance.GetDefinition(DefinitionsCategory.MISSION_TYPES, missionDef.Get("type")),
-			_data["targetValue"].AsFloat, 
+			_data["targetValue"].AsLong, 
 			_data["singleRun"].AsBool
 		);
 
@@ -309,7 +309,7 @@ public class Mission {
 
 		// Restore objective
 		if(m_objective != null) {
-			m_objective.tracker.SetValue(_data["currentValue"].AsFloat, false);
+			m_objective.tracker.SetValue(_data["currentValue"].AsLong, false);
 			m_objective.enabled = (m_state == State.ACTIVE);
 		}
